moves DBError up a folder and updates readme:
